+++ include/os_malloc.h
@@ -1,0 +1,53 @@
+#ifndef OS_MALLOC_H
+#define OS_MALLOC_H
+
+#include "ultra64.h"
+
+struct ArenaNode;
+
+typedef struct Arena {
+    /* 0x00 */ struct ArenaNode* head;
+    /* 0x04 */ void* start;
+    /* 0x08 */ OSMesgQueue lockQueue;
+    /* 0x20 */ u8 unk_20;
+    /* 0x21 */ u8 isInit;
+    /* 0x22 */ u8 flag;
+} Arena; // size = 0x24
+
+typedef struct ArenaNode {
+    /* 0x00 */ s16 magic;
+    /* 0x02 */ s16 isFree;
+    /* 0x04 */ u32 size;
+    /* 0x08 */ struct ArenaNode* next;
+    /* 0x0C */ struct ArenaNode* prev;
+    /* 0x10 */ const char* filename;
+    /* 0x14 */ s32 line;
+    /* 0x18 */ OSId threadId;
+    /* 0x1C */ Arena* arena;
+    /* 0x20 */ OSTime time;
+    /* 0x28 */ u8 unk_28[0x30-0x28]; // probably padding
+} ArenaNode; // size = 0x30
+
+void __osMallocInit(Arena* arena, void* start, u32 size);
+void __osMallocCleanup(Arena* arena);
+u8 __osMallocIsInitialized(Arena* arena);
+void* __osMalloc_NoLockDebug(Arena* arena, u32 size, const char* file, s32 line);
+void* __osMallocDebug(Arena* arena, u32 size, const char* file, s32 line);
+void* __osMallocRDebug(Arena* arena, u32 size, const char* file, s32 line);
+void* __osMalloc_NoLock(Arena* arena, u32 size);
+void* __osMalloc(Arena* arena, u32 size);
+void* __osMallocR(Arena* arena, u32 size);
+void __osFree_NoLock(Arena* arena, void* ptr);
+void __osFree(Arena* arena, void* ptr);
+void __osFree_NoLockDebug(Arena* arena, void* ptr, const char* file, s32 line);
+void __osFreeDebug(Arena* arena, void* ptr, const char* file, s32 line);
+void* __osRealloc(Arena* arena, void* ptr, u32 newSize);
+void* __osReallocDebug(Arena* arena, void* ptr, u32 newSize, const char* file, s32 line);
+void ArenaImpl_GetSizes(Arena* arena, u32* outMaxFree, u32* outFree, u32* outAlloc);
+void __osDisplayArena(Arena* arena);
+u32 __osCheckArena(Arena* arena);
+u8 func_800FF334(Arena* arena);
+
+extern u32 __osMalloc_FreeBlockTest_Enable;
+
+#endif
