+++ include/variables.h
@@ -217,8 +217,7 @@ extern ActiveSequence gActiveSeqs[4];
 extern AudioContext gAudioCtx;
 extern void(*D_801755D0)(void);
 
-extern u32 __osMalloc_FreeBlockTest_Enable;
-extern Arena gSystemArena;
+extern struct Arena gSystemArena;
 extern OSPifRam __osContPifRam;
 extern u8 __osContLastCmd;
 extern u8 __osMaxControllers;
