+++ src/code/z_malloc.c
@@ -1,4 +1,5 @@
 #include "global.h"
+#include "os_malloc.h"
 
 #define LOG_SEVERITY_NOLOG 0
 #define LOG_SEVERITY_ERROR 2
